Move Parts into it's own custom element
